% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciclo_mareal.R
\name{ciclo_mareal}
\alias{ciclo_mareal}
\title{Grafica del ciclo mareal.}
\usage{
ciclo_mareal(datos, fecha_hora, altura)
}
\arguments{
\item{datos:}{Conjunto de datos de donde se obtiene el ciclo mareal.}

\item{fecha_hora:}{Vector o columna con los datos de la fecha y la hora guardados como as.POSIXct(). Se debe usar el pacquete "scales"}

\item{altura:}{vector o comna con los datos de la Altura mareal referída al MLWS (m).}
}
\description{
Grafica del ciclo mareal.
}
