% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfil_en_profundidad.R
\name{perfil_en_profundidad}
\alias{perfil_en_profundidad}
\title{Title}
\usage{
perfil_en_profundidad(
  Estacion,
  Titulo,
  VarProfundidad,
  variable,
  labelx,
  labely,
  profundidadMax
)
}
\arguments{
\item{profundidadMax}{}
}
\description{
Title
}
